pyproject.toml
[project]
name = "company-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.19.0",
    "gradio>=5.23.1",
    "mlflow>=3.6.0",
    "pydantic>=2.12.5",
]
